reorganized files and using howto notebooks for testing
